# [SPARK-53239][SQL] Improve `MapSort` and `SortArray` performance via `parallelSort`