### What changes were proposed in this pull request?

This PR aims to improve `MapSort` and `SortArray` performance by switching to `Arrays.parallelSort` from `Arrays.sort` to utilize the multi-core more if applicable.

```scala
// Large data: 10M
scala> val a = scala.util.Random.shuffle(1 to 10_000_000).toArray
scala> val b = a.clone
scala> spark.time(java.util.Arrays.sort(a))
Time taken: 574 ms
scala> spark.time(java.util.Arrays.parallelSort(b))
Time taken: 160 ms

// Small data
scala> val a = scala.util.Random.shuffle(1 to 10).toArray
scala> val b = a.clone()
scala> spark.time(java.util.Arrays.sort(a))
Time taken: 0 ms
scala> spark.time(java.util.Arrays.parallelSort(b))
Time taken: 0 ms
```

### Why are the changes needed?

**DATA GENERATION**

```scala
scala> sql("CREATE TABLE T AS SELECT shuffle(sequence(1, 10000000)) AS a")
```

**AFTER (4s)**

```scala

scala> spark.time(sql("SELECT sort_array(a) FROM T").show())
+--------------------+
| sort_array(a, true)|
+--------------------+
|[1, 2, 3, 4, 5, 6...|
+--------------------+

Time taken: 4320 ms

scala> spark.time(sql("SELECT sort_array(a) FROM T").show())
+--------------------+
| sort_array(a, true)|
+--------------------+
|[1, 2, 3, 4, 5, 6...|
+--------------------+

Time taken: 3903 ms
```

**BEFORE (4.1.0-preview1): 6s**

```scala
scala> spark.version
val res0: String = 4.1.0-preview1

scala> spark.time(sql("SELECT sort_array(a) FROM T").show())
+--------------------+
| sort_array(a, true)|
+--------------------+
|[1, 2, 3, 4, 5, 6...|
+--------------------+

Time taken: 8774 ms

scala> spark.time(sql("SELECT sort_array(a) FROM T").show())
+--------------------+
| sort_array(a, true)|
+--------------------+
|[1, 2, 3, 4, 5, 6...|
+--------------------+

Time taken: 6043 ms
```

### Does this PR introduce _any_ user-facing change?

No behavior change.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.